#4 CH05 리팩토링

+++ src/main/java/com/study/objects/_05_movie/PercentDiscountMovie.java
@@ -1,18 +1,19 @@
 package com.study.objects._05_movie;
 
+import java.math.BigDecimal;
 import java.time.Duration;
 import java.util.List;
 
 public class PercentDiscountMovie extends Movie {
 
-  private final double discountPercent;
+  private final BigDecimal discountPercent;
 
   public PercentDiscountMovie(
       String title,
       Duration runningTime,
       Money fee,
       List<DiscountCondition> conditions,
-      double discountPercent,
+      BigDecimal discountPercent,
       MovieType type
   ) {
     super(title, runningTime, fee, conditions, type); // 부모 생성자 호출
